+++ test/unit/util.test.ts
@@ -268,8 +268,6 @@ describe('util.test.js', () => {
         it('should work with non latin-1 chars', async () => {
             const plain = 'aäß';
             const base64 = 'YcOkw58=';
-
-            console.log('-----------------');
             const blobBuffer = blobBufferUtil.createBlobBuffer(plain, 'plain/text');
             assert.strictEqual(
                 await blobBufferUtil.toString(blobBuffer),